refactor: convertMessagesを別ファイルに分離し循環依存を解消
- mlx-cache-controller.ts → mlx-driver.ts の循環importを解消
- convertMessagesをmlx-message-utils.tsに切り出し
- キャッシュファイル名にフルSHA-256ハッシュを使用（衝突防止）
