+++ packages/driver/src/mlx-ml/mlx-driver.ts
@@ -1,18 +1,18 @@
 import { Readable } from 'stream';
 import type { AIDriver, QueryOptions, QueryResult, StreamResult, ToolDefinition, FinishReason } from '../types.js';
-import { hasToolCalls, isToolResult } from '../types.js';
-import type { FormatterOptions, ChatMessage } from '../formatter/types.js';
+import { isToolResult } from '../types.js';
+import type { FormatterOptions } from '../formatter/types.js';
 import { formatPromptAsMessages } from '../formatter/converter.js';
 import { formatCompletionPrompt } from '../formatter/completion-formatter.js';
 import { MlxProcess } from './process/index.js';
-import type { MlxMessage, MlxMlModelOptions, MlxModelCapabilities, MlxContentPart } from './types.js';
+import type { MlxMlModelOptions, MlxModelCapabilities } from './types.js';
 import type { MlxRuntimeInfo, MlxToolDefinition } from './process/types.js';
 import { createModelSpecificProcessor, selectApi } from './process/model-specific.js';
 import { selectResponseProcessor } from './process/model-handlers.js';
 import type { CompiledPrompt } from '@modular-prompt/core';
 import { extractJSON } from '@modular-prompt/utils';
 import { formatToolDefinitionsAsText } from './tool-call-parser.js';
-import { contentToString, extractImagePaths } from '../content-utils.js';
+import { convertMessages, extractImagePaths } from './mlx-message-utils.js';
 import { QueryLogger } from '../query-logger.js';
 import type { PromptCacheController } from '../cache-controller.js';
 import { extractCacheablePrefix } from '../cache-utils.js';
@@ -51,67 +51,6 @@ export function hasMessageElement(prompt: CompiledPrompt): boolean {
   return checkElements(prompt.instructions) ||
     checkElements(prompt.data) ||
     checkElements(prompt.output);
-}
-
-/**
- * Convert ChatMessage format to MLX format
- * VLM mode preserves image placeholders as structured content parts
- */
-export function convertMessages(messages: ChatMessage[], vlm: boolean = false): MlxMessage[] {
-  return messages.map(msg => {
-    // AssistantToolCallMessage - tool_calls付きメッセージ
-    if (hasToolCalls(msg)) {
-      return {
-        role: 'assistant' as const,
-        content: msg.content,
-        tool_calls: msg.toolCalls.map(tc => ({
-          id: tc.id,
-          type: 'function' as const,
-          function: {
-            name: tc.name,
-            arguments: JSON.stringify(tc.arguments)
-          }
-        }))
-      };
-    }
-
-    // ToolResultMessage - ツール結果メッセージ
-    if (isToolResult(msg)) {
-      let content: string;
-      if (msg.kind === 'text') {
-        content = String(msg.value);
-      } else if (msg.kind === 'data') {
-        content = JSON.stringify(msg.value);
-      } else {
-        content = String(msg.value);
-      }
-      return {
-        role: 'tool' as const,
-        content,
-        tool_call_id: msg.toolCallId,
-        name: msg.name
-      };
-    }
-
-    // StandardChatMessage - 通常メッセージ（VLM対応含む）
-    if (vlm && Array.isArray(msg.content)) {
-      const parts: MlxContentPart[] = [];
-      for (const att of msg.content) {
-        if (att.type === 'image_url' && att.image_url?.url) {
-          parts.push({ type: 'image' });
-        } else if (att.type === 'text' && att.text) {
-          parts.push({ type: 'text', text: att.text });
-        }
-      }
-      if (parts.length > 0) {
-        return { role: msg.role as 'system' | 'user' | 'assistant', content: parts };
-      }
-    }
-    return {
-      role: msg.role as 'system' | 'user' | 'assistant',
-      content: contentToString(msg.content)
-    };
-  });
 }
 
 // ========================================================================
